% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addb-package.R
\docType{data}
\name{aus.fertility}
\alias{aus.fertility}
\title{Australian fertility data}
\format{Object of class \code{demogdata} containing the following
components:
\describe{
  \item{year}{Vector of years}
  \item{age}{Vector of ages}
  \item{rate}{List containing one matrix with one age group per row and
              one column per year.}
  \item{pop}{Population data in same form as \code{rate}.}
  \item{type}{Type of object. In this case, \dQuote{fertility}.}
  \item{label}{Character string giving area from which data are taken. In this
               case, \dQuote{Australia}.}
}}
\description{
Age-specific fertility rates and female child-bearing population for
Australia.
}
\details{
Australian fertility rates and populations (1921-2006) for single year age
groups (15-49) as provided by the Australian Bureau of Statistics. The
population numbers for 2002 onwards differ slighly from those in the
\code{\link{australia}} (mortality) object. The population numbers in
\code{aus.fertility} have been updated to the latest ABS estimates.
}
\examples{

plot(aus.fertility)

}
\seealso{
\code{\link[demography]{aus.fert}} in the
\code{\link[demography]{demography}} package contains similar data in 5-year
age groups.
}
\author{
Rob J Hyndman
}
\keyword{data}
