% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addb-package.R
\docType{data}
\name{data}
\alias{data}
\alias{australia}
\alias{nsw}
\alias{vic}
\alias{qld}
\alias{tas}
\alias{sa}
\alias{nt}
\alias{wa}
\alias{actot}
\alias{addb}
\title{Mortality and population data}
\format{All data are \code{demogdata} objects containing the following
components:
  \describe{
    \item{year}{Years for which data available}
    \item{age}{Ages for which data available}
    \item{rates}{A list with matrix components \code{female}, \code{male} and \code{total} with one age group
                 per row and one column per year.}
    \item{pop}{A list of the same form as \code{rates} but containing population numbers instead of mortality rates.}
    \item{type}{A character string equal to \code{"mortality"} in each case.}
    \item{label}{A character string giving the geographical region.}
}}
\source{
Smith, L. (2009)
The Australian Demographic Data Bank, 1901-2003: Populations [Computer File].
Canberra: Australian Social Science Data Archive,
The Australian National University, 2009.
Smith, L. (2009)
The Australian Demographic Data Bank, 1901-2003: Deaths [Computer File].
Canberra: Australian Social Science Data Archive,
The Australian National University, 2009.
}
\description{
Mortality rates and populations for Australia and Australian states. The
data can be plotted and modelled using the functions in the demography
package such as \code{\link[demography]{plot.demogdata}} and
\code{\link[demography]{fdm}}.
}
\details{
The following data sets are included:
  \tabular{lll}{
    \code{australia} \tab Australia \tab (1901-2003)\cr
    \code{nsw} \tab New South Wales \tab (1901-2003)\cr
    \code{vic} \tab Victoria \tab (1901-2003)\cr
    \code{qld} \tab Queensland \tab (1901-2003)\cr
    \code{sa} \tab South Australia \tab (1901-2003)\cr
    \code{wa} \tab Western Australia \tab (1901-2003)\cr
    \code{nt} \tab Northern Territory \tab (1901-2003)\cr
    \code{actot} \tab Australian Capital Territory and overseas territories \tab (1901-2003)\cr
    \code{tas} \tab Tasmania \tab (1901-2003)
}
All data are from v3.2b of the
Australian Demographic Data Bank released 10 February 2005.
}
\examples{

plot(vic,years=1950:1997)

}
\author{
Rob J Hyndman
}
\keyword{data}
